functions defining fastapi creation moved from main to src/app/core/setup.py, main now imports and creates only
