+++ src/app/core/setup.py
@@ -1,0 +1,158 @@
+from fastapi import FastAPI, APIRouter, Depends
+from fastapi.openapi.docs import get_swagger_ui_html, get_redoc_html
+from fastapi.openapi.utils import get_openapi
+import redis.asyncio as redis
+from arq import create_pool
+from arq.connections import RedisSettings
+import anyio
+
+
+from app.api.dependencies import get_current_superuser
+from app.core import cache, queue
+from app.core.config import settings
+from app.core.database import Base
+from app.core.database import async_engine as engine
+from app.core.config import (
+    DatabaseSettings, 
+    RedisCacheSettings, 
+    AppSettings, 
+    ClientSideCacheSettings, 
+    RedisQueueSettings,
+    EnvironmentOption,
+    EnvironmentSettings
+)
+
+# -------------- database --------------
+async def create_tables():
+    async with engine.begin() as conn:
+        await conn.run_sync(Base.metadata.create_all)
+
+
+# -------------- cache --------------
+async def create_redis_cache_pool():
+    cache.pool = redis.ConnectionPool.from_url(settings.REDIS_CACHE_URL)
+    cache.client = redis.Redis.from_pool(cache.pool)
+
+
+async def close_redis_cache_pool():
+    await cache.client.aclose()
+
+
+# -------------- queue --------------
+async def create_redis_queue_pool():
+    queue.pool = await create_pool(
+        RedisSettings(host=settings.REDIS_QUEUE_HOST, port=settings.REDIS_QUEUE_PORT)
+    )
+
+
+async def close_redis_queue_pool():
+    await queue.pool.aclose()
+
+
+# -------------- application --------------
+async def set_threadpool_tokens(number_of_tokens=100):
+    limiter = anyio.to_thread.current_default_thread_limiter()
+    limiter.total_tokens = number_of_tokens
+
+
+# -------------- application --------------
+def create_application(router: APIRouter, settings, **kwargs) -> FastAPI:
+    """
+    Creates and configures a FastAPI application based on the provided settings.
+
+    This function initializes a FastAPI application, then conditionally configures 
+    it with various settings and handlers. The specific configuration is determined 
+    by the type of the `settings` object provided.
+
+    Parameters
+    ----------
+    router : APIRouter
+        The APIRouter object that contains the routes to be included in the FastAPI application.
+
+    settings
+        An instance representing the settings for configuring the FastAPI application. It determines the configuration applied:
+
+        - AppSettings: Configures basic app metadata like name, description, contact, and license info.
+        - DatabaseSettings: Adds event handlers for initializing database tables during startup.
+        - RedisCacheSettings: Sets up event handlers for creating and closing a Redis cache pool.
+        - ClientSideCacheSettings: Integrates middleware for client-side caching.
+        - RedisQueueSettings: Sets up event handlers for creating and closing a Redis queue pool.
+        - EnvironmentSettings: Conditionally sets documentation URLs and integrates custom routes for API documentation based on environment type.
+
+    **kwargs 
+        Extra keyword arguments passed directly to the FastAPI constructor.
+
+    Returns
+    -------
+    FastAPI
+        A fully configured FastAPI application instance.
+
+    """
+
+    # --- before creating application ---
+    if isinstance(settings, AppSettings):
+        to_update = {
+            "title": settings.APP_NAME,
+            "description": settings.APP_DESCRIPTION,
+            "contact": {
+                "name": settings.CONTACT_NAME,
+                "email": settings.CONTACT_EMAIL
+            },
+            "license_info": {
+                "name": settings.LICENSE_NAME
+            }
+        }
+        kwargs.update(to_update)
+
+    if isinstance(settings, EnvironmentSettings):
+        kwargs.update(
+            {
+                "docs_url": None, 
+                "redoc_url": None, 
+                "openapi_url": None
+            }
+        )
+
+    application = FastAPI(**kwargs)
+
+    # --- application created ---
+    application.include_router(router)
+    application.add_event_handler("startup", set_threadpool_tokens)
+
+    if isinstance(settings, DatabaseSettings):
+        application.add_event_handler("startup", create_tables)
+    
+    if isinstance(settings, RedisCacheSettings):
+        application.add_event_handler("startup", create_redis_cache_pool)
+        application.add_event_handler("shutdown", close_redis_cache_pool)
+
+    if isinstance(settings, ClientSideCacheSettings):
+        application.add_middleware(cache.ClientCacheMiddleware, max_age=60)
+
+    if isinstance(settings, RedisQueueSettings):
+        application.add_event_handler("startup", create_redis_queue_pool)
+        application.add_event_handler("shutdown", close_redis_queue_pool)
+
+    if isinstance(settings, EnvironmentSettings):
+        if settings.ENVIRONMENT != EnvironmentOption.PRODUCTION:
+            docs_router = APIRouter()
+            if settings.ENVIRONMENT != EnvironmentOption.LOCAL:
+                docs_router = APIRouter(dependencies=[Depends(get_current_superuser)])
+            
+            @docs_router.get("/docs", include_in_schema=False)
+            async def get_swagger_documentation():
+                return get_swagger_ui_html(openapi_url="/openapi.json", title="docs")
+
+
+            @docs_router.get("/redoc", include_in_schema=False)
+            async def get_redoc_documentation():
+                return get_redoc_html(openapi_url="/openapi.json", title="docs")
+
+
+            @docs_router.get("/openapi.json", include_in_schema=False)
+            async def openapi():
+                return get_openapi(title=application.title, version=application.version, routes=application.routes)
+            
+            application.include_router(docs_router)
+            
+    return application
